Small refactoring to keep main homesick class smaller.
